+++ schema_graph/views.py
@@ -11,8 +11,16 @@ class Schema(TemplateView):
     template_name = "schema_graph/schema.html"
 
     def access_permitted(self):
-        """When this returns True, the schema graph page is accessible."""
-        return settings.DEBUG
+        """
+        When this returns True, the schema graph page is accessible.
+
+        We look for the setting `SCHEMA_GRAPH_VISIBLE`, and fall back to `DEBUG`.
+
+        To control this on a per-request basis, override this function in a subclass.
+        The request will be accessible using `self.request`.
+        """
+
+        return getattr(settings, "SCHEMA_GRAPH_VISIBLE", settings.DEBUG)
 
     def dispatch(self, request):
         if not self.access_permitted():